requirements.txt
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.0
pip-chill==1.0.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
